+++ CHANGELOG.md
@@ -25,6 +25,7 @@ in the detailed section referring to by linking pull requests or issues.
 * CloudEvents Http extension (#1160)
 * Full Query capabilities for SQL TransferProcess Store (#1483)
 * Full Query capabilities for SQL PolicyDefinition Store (#1484)
+* Allow `TypeManager` to support multiple serialization contexts (#1581)
 
 #### Changed